Move payment stuff to the invoicing package
