+++ service/src/main/kotlin/fi/espoo/evaka/invoicing/data/PaymentQueries.kt
@@ -2,8 +2,14 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-package fi.espoo.evaka.payments
+package fi.espoo.evaka.invoicing.data
 
+import fi.espoo.evaka.invoicing.controller.PaymentDistinctiveParams
+import fi.espoo.evaka.invoicing.controller.PaymentSortParam
+import fi.espoo.evaka.invoicing.controller.SearchPaymentsRequest
+import fi.espoo.evaka.invoicing.domain.Payment
+import fi.espoo.evaka.invoicing.domain.PaymentDraft
+import fi.espoo.evaka.invoicing.domain.PaymentUnit
 import fi.espoo.evaka.shared.DaycareId
 import fi.espoo.evaka.shared.Paged
 import fi.espoo.evaka.shared.PaymentId
