+++ service/src/main/kotlin/fi/espoo/evaka/invoicing/service/PaymentService.kt
@@ -2,16 +2,25 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-package fi.espoo.evaka.payments
+package fi.espoo.evaka.invoicing.service
 
+import fi.espoo.evaka.invoicing.data.getMaxPaymentNumber
+import fi.espoo.evaka.invoicing.data.getPaymentUnitsByIds
+import fi.espoo.evaka.invoicing.data.readPaymentsByIds
+import fi.espoo.evaka.invoicing.data.updatePaymentDraftsAsSent
+import fi.espoo.evaka.invoicing.domain.PaymentIntegrationClient
+import fi.espoo.evaka.invoicing.domain.PaymentStatus
 import fi.espoo.evaka.shared.FeatureConfig
 import fi.espoo.evaka.shared.PaymentId
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.domain.BadRequest
 import fi.espoo.evaka.shared.domain.HelsinkiDateTime
+import mu.KotlinLogging
 import org.springframework.stereotype.Component
 import java.time.LocalDate
+
+private val logger = KotlinLogging.logger {}
 
 @Component
 class PaymentService(
